% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualificacao}
\alias{qualificacao}
\title{Partners participation type description}
\format{A data frame with 40 rows and 2 variables:
\describe{
  \item{qualificacao}{Code to be joined with company partners data}
  \item{qualificacao_nm}{Description of participation type}
  ...
}}
\source{
\url{http://idg.receita.fazenda.gov.br/orientacao/tributaria/cadastros/cadastro-nacional-de-pessoas-juridicas-cnpj/Qualificacao_socio.pdf}
}
\usage{
qualificacao
}
\description{
A dataset description of participation types in RFB partner (socios) data.
}
\keyword{datasets}
