% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{rfb_import}
\alias{rfb_import}
\title{Import data from binary files}
\usage{
rfb_import(type = c("all", "empresas", "socios"))
}
\arguments{
\item{type}{if \code{type="all"}, download list-column tibble containing all data. If \code{type="empresas"}, downloads rectangular database of companies. If \code{type="socios"}, downloads rectangular database of partners.}
}
\description{
Downloads and reads data directly from Kaggle Datasets, where we have uploaded parsed data.
}
\examples{
\dontrun{
empresas <- rfb_import("empresas")
}
}
