% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{rfb_read}
\alias{rfb_read}
\alias{rfb_read_dir}
\title{Read CNPJ files}
\usage{
rfb_read(file_name, remove_temp_files = TRUE)

rfb_read_dir(dir_name, remove_temp_files = TRUE)
}
\arguments{
\item{file_name}{character string containing full paths of the files.}

\item{remove_temp_files}{remove temporary fixed width files}

\item{dir_name}{path to directory where text files are}
}
\value{
list-column tibble containing data of companies and partners
}
\description{
Read CNPJ files from previously downloaded fixed width files, obtained
  from \code{\link{rfb_download}}

Lists all files in directory and read them
}
